// src/components/CardWrapper/types.ts
export type VariantTypoRounded = '0' | '8' | '24' | '40' | '16' | '12' | '100';

export type VariantTypoStoke = '1' | '2';

export type VariantTypoColors =
  | 'black_1'
  | 'black_2'
  | 'black_3'
  | 'black_4'
  | 'grey_1'
  | 'grey_2'
  | 'grey_3'
  | 'grey_4'
  | 'white_1'
  | 'white_2'
  | 'white_3'
  | 'white_4'
  | 'white'
  | 'pr_purple'
  | 'sec_purple'
  | 'pr_blue'
  | 'sec_blue'
  | 'error'
  | 'warning'
  | 'success'
  | 'backdrop_2';